third_party/WebKit/LayoutTests/platform/win/bluetooth/characteristic/notifications/characteristic-is-removed-expected.txt

This is a testharness.js-based test.
PASS Characteristic is removed. Reject with InvalidStateError. 
Harness: the test ran to completion.
